+++ tapo-py/src/requests.rs
@@ -1,5 +1,7 @@
 mod energy_data_interval;
+mod play_alarm;
 mod set_device_info;
 
 pub use energy_data_interval::*;
+pub use play_alarm::*;
 pub use set_device_info::*;