+++ tapo-py/src/requests/play_alarm.rs
@@ -1,0 +1,9 @@
+use pyo3::prelude::*;
+
+#[derive(Debug, Clone, PartialEq)]
+#[pyclass(name = "AlarmDuration", eq)]
+pub enum PyAlarmDuration {
+    Continuous,
+    Once,
+    Seconds,
+}