+++ tapo/examples/tapo_h100.rs
@@ -115,20 +115,19 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         }
     }
 
-    let ringtone = AlarmRingtone::Alarm1;
-    let volume = AlarmVolume::Low;
-    let duration = AlarmDuration::Seconds(1);
-
-    info!("Triggering the alarm ringtone {ringtone:?} for {duration:?} at a {volume:?} volume");
-    hub.play_alarm(Some(ringtone), Some(volume), duration)
-        .await?;
+    info!("Triggering the alarm ringtone 'Alarm 1' at a 'Low' volume for '3 Seconds'...");
+    hub.play_alarm(
+        AlarmRingtone::Alarm1,
+        AlarmVolume::Low,
+        AlarmDuration::Seconds(3),
+    )
+    .await?;
 
     let device_info = hub.get_device_info().await?;
     info!("Is device ringing?: {:?}", device_info.in_alarm);
 
+    info!("Stopping the alarm after 1 Second...");
     tokio::time::sleep(Duration::from_secs(1)).await;
-
-    info!("Stopping the alarm");
     hub.stop_alarm().await?;
 
     let device_info = hub.get_device_info().await?;