+++ tapo/src/api/hub_handler.rs
@@ -58,38 +58,6 @@ impl HubHandler {
         self.client.read().await.get_device_info().await
     }
 
-    /// Returns a list of ringtones (alarm types) supported by the hub.
-    /// Used for debugging only.
-    pub async fn get_supported_ringtone_list(&self) -> Result<Vec<String>, Error> {
-        self.client
-            .read()
-            .await
-            .get_supported_alarm_type_list()
-            .await
-            .map(|response| response.alarm_type_list)
-    }
-
-    /// Start playing the hub alarm.
-    /// By default, this uses the configured alarm settings on the hub.
-    /// Each of the settings can be overridden by passing `Some` as a parameter.
-    pub async fn play_alarm(
-        &self,
-        ringtone: Option<AlarmRingtone>,
-        volume: Option<AlarmVolume>,
-        duration: AlarmDuration,
-    ) -> Result<(), Error> {
-        self.client
-            .read()
-            .await
-            .play_alarm(PlayAlarmParams::new(ringtone, volume, duration)?)
-            .await
-    }
-
-    /// Stop playing the hub alarm if currently playing
-    pub async fn stop_alarm(&self) -> Result<(), Error> {
-        self.client.read().await.stop_alarm().await
-    }
-
     /// Returns *device info* as [`serde_json::Value`].
     /// It contains all the properties returned from the Tapo API.
     pub async fn get_device_info_json(&self) -> Result<serde_json::Value, Error> {
@@ -148,6 +116,36 @@ impl HubHandler {
             .await
             .get_child_device_component_list()
             .await
+    }
+
+    /// Returns a list of ringtones (alarm types) supported by the hub.
+    /// Used for debugging only.
+    pub async fn get_supported_ringtone_list(&self) -> Result<Vec<String>, Error> {
+        self.client
+            .read()
+            .await
+            .get_supported_alarm_type_list()
+            .await
+            .map(|response| response.alarm_type_list)
+    }
+
+    /// Start playing the hub alarm.
+    pub async fn play_alarm(
+        &self,
+        ringtone: AlarmRingtone,
+        volume: AlarmVolume,
+        duration: AlarmDuration,
+    ) -> Result<(), Error> {
+        self.client
+            .read()
+            .await
+            .play_alarm(PlayAlarmParams::new(ringtone, volume, duration)?)
+            .await
+    }
+
+    /// Stop playing the hub alarm, if it's currently playing.
+    pub async fn stop_alarm(&self) -> Result<(), Error> {
+        self.client.read().await.stop_alarm().await
     }
 }